+++ hudi-spark-datasource/hudi-spark/src/test/java/HoodieJavaGenerateApp.java
@@ -24,9 +24,9 @@ import org.apache.hudi.common.table.timeline.HoodieActiveTimeline;
 import org.apache.hudi.common.testutils.HoodieTestDataGenerator;
 import org.apache.hudi.config.HoodieWriteConfig;
 import org.apache.hudi.hive.HiveSyncConfig;
-import org.apache.hudi.hive.MultiPartKeysValueExtractor;
-import org.apache.hudi.hive.NonPartitionedExtractor;
-import org.apache.hudi.hive.SlashEncodedDayPartitionValueExtractor;
+import org.apache.hudi.sync.common.model.partextractor.MultiPartKeysValueExtractor;
+import org.apache.hudi.sync.common.model.partextractor.NonPartitionedExtractor;
+import org.apache.hudi.sync.common.model.partextractor.SlashEncodedDayPartitionValueExtractor;
 import org.apache.hudi.keygen.NonpartitionedKeyGenerator;
 import org.apache.hudi.keygen.SimpleKeyGenerator;
