+++ hudi-sync/hudi-sync-common/src/main/java/org/apache/hudi/sync/common/model/partextractor/MultiPartKeysValueExtractor.java
@@ -7,20 +7,20 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
-package org.apache.hudi.hive;
+package org.apache.hudi.sync.common.model.partextractor;
 
 import java.util.Collections;
 import org.apache.hudi.common.util.ValidationUtils;
-import org.apache.hudi.sync.common.model.PartitionValueExtractor;
 
 import java.util.Arrays;
 import java.util.List;
